# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Silk Way API",
        "description": "REST API для мобильного приложения Silk Way. Авторизация через Bearer токен (Sanctum).",
        "contact": {
            "email": "admin@silkway.kz"
        },
        "version": "1.9.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "c47a1b563e5e8bc3c7b962f893981e84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "warehouse_employee",
                                        "enum": [
                                            "warehouse_employee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Регистрация успешна, ожидайте подтверждения администратора",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Регистрация успешна. Ожидайте подтверждения администратора."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вход в систему",
                "operationId": "1ff9c65c5cd1a8e0786e318eed856972",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учётные данные"
                    },
                    "403": {
                        "description": "Аккаунт не подтверждён"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход из системы (удаляет текущий токен)",
                "operationId": "eae6a8251a6469bbabe454860f64cf37",
                "responses": {
                    "200": {
                        "description": "Выход выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Выход выполнен."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Данные текущего пользователя",
                "operationId": "257a57ee6760f86fb4b11ff94dba1f38",
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/auth/driver/login/request": {
            "post": {
                "tags": [
                    "DriverLogin"
                ],
                "summary": "Шаг 1 — Запрос OTP-кода для входа через WhatsApp",
                "description": "Находит водителя по номеру телефона, проверяет статус аккаунта и отправляет 6-значный код в WhatsApp. Повторная отправка — не чаще 1 раза в 60 секунд.",
                "operationId": "fe1cefae8e14e206b1595a655071795d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+7 700 120 00 13"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Код отправлен в WhatsApp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Код для входа отправлен"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "77001200013"
                                        },
                                        "phone_masked": {
                                            "type": "string",
                                            "example": "+7 ••• ••• **-13"
                                        },
                                        "expires_in_seconds": {
                                            "type": "integer",
                                            "example": 600
                                        },
                                        "resend_available_in_seconds": {
                                            "type": "integer",
                                            "example": 60
                                        },
                                        "register_url": {
                                            "type": "string",
                                            "example": "https://example.com/register/driver"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Аккаунт ещё не подтверждён администратором"
                    },
                    "404": {
                        "description": "Водитель с таким номером не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "429": {
                        "description": "Слишком часто — подождите 60 секунд"
                    },
                    "503": {
                        "description": "WAHA-сессия не готова"
                    }
                }
            }
        },
...
        "/auth/driver/register/request": {
            "post": {
                "tags": [
                    "DriverRegistration"
                ],
                "summary": "Шаг 1 — Запрос OTP-кода через WhatsApp",
                "description": "Проверяет номер телефона, отправляет 6-значный код в WhatsApp. Повторная отправка — не чаще 1 раза в 60 секунд.",
                "operationId": "aa15dc2aade27ef6ecf03174c965691b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иван Иванов",
                                        "maxLength": 255
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+7 700 120 00 13"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Код отправлен в WhatsApp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Код отправлен"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "77001200013"
                                        },
                                        "expires_in_seconds": {
                                            "type": "integer",
                                            "example": 600
                                        },
                                        "resend_available_in_seconds": {
                                            "type": "integer",
                                            "example": 60
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Номер уже зарегистрирован"
                    },
                    "422": {
                        "description": "Ошибка валидации / номер не найден в WhatsApp"
                    },
                    "429": {
                        "description": "Слишком часто — подождите 60 секунд"
                    },
                    "503": {
                        "description": "WAHA-сессия не готова"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "example": "ivan@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "warehouse_employee",
                            "driver"
                        ]
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введите токен (без слова Bearer — оно добавляется автоматически)",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}